+++ covilo-core/src/main/kotlin/me/kvdpxne/covilo/domain/persistence/CrimePerpetratorRepository.kt
@@ -10,13 +10,16 @@ import java.util.UUID
 interface CrimePerpetratorRepository {
 
   fun findByIdentifier(identifier: UUID): CrimePerpetrator?
+
   fun findAll(): CrimePerpetrators
 
   fun insert(perpetrator: CrimePerpetrator)
+
   fun update(perpetrator: CrimePerpetrator)
 
   fun delete(identifier: UUID)
+
   fun deleteAll()
 
-  fun count(): Long
+  fun count(): Int
 }
